+++ libpod/lock/shm/shm_lock_test.go
@@ -1,4 +1,4 @@
-package lock
+package shm
 
 import (
 	"fmt"
